requirements.txt
llama-index
nltk
milvus==2.2.14
pymilvus==2.3.3
langchain
python-dotenv
openai
flask
pypdf
rapidocr-onnxruntime
chromadb
lark

